+++ pac4j-oauth/src/main/java/org/pac4j/oauth/client/FoursquareClient.java
@@ -1,17 +1,10 @@
 package org.pac4j.oauth.client;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import com.github.scribejava.apis.Foursquare2Api;
-import com.github.scribejava.core.builder.api.BaseApi;
-import com.github.scribejava.core.builder.api.DefaultApi20;
-import com.github.scribejava.core.model.OAuth2AccessToken;
-import com.github.scribejava.core.oauth.OAuth20Service;
 import org.pac4j.core.context.WebContext;
-import org.pac4j.core.exception.HttpAction;
-import org.pac4j.oauth.profile.JsonHelper;
 import org.pac4j.oauth.profile.foursquare.FoursquareProfile;
+import org.pac4j.oauth.profile.foursquare.FoursquareProfileCreator;
 import org.pac4j.oauth.profile.foursquare.FoursquareProfileDefinition;
-import org.pac4j.scribe.oauth.Foursquare20Service;
 
 /**
  * <p>This class is the OAuth client to authenticate users in Foursquare.
@@ -21,7 +14,7 @@ import org.pac4j.scribe.oauth.Foursquare20Service;
  * @author Alexey Ogarkov
  * @since 1.5.0
  */
-public class FoursquareClient extends BaseOAuth20Client<FoursquareProfile>{
+public class FoursquareClient extends OAuth20Client<FoursquareProfile>{
 
     public FoursquareClient() {}
 
@@ -32,45 +25,12 @@ public class FoursquareClient extends BaseOAuth20Client<FoursquareProfile>{
 
     @Override
     protected void internalInit(final WebContext context) {
+        configuration.setApi(Foursquare2Api.instance());
+        configuration.setProfileDefinition(new FoursquareProfileDefinition());
+        configuration.setScope("user");
+        setConfiguration(configuration);
+        setProfileCreator(new FoursquareProfileCreator(configuration));
+
         super.internalInit(context);
-        this.service = new Foursquare20Service((DefaultApi20) getApi(), buildOAuthConfig(context));
-        setProfileDefinition(new FoursquareProfileDefinition());
-    }
-
-    @Override
-    protected BaseApi<OAuth20Service> getApi() {
-        return Foursquare2Api.instance();
-    }
-
-    @Override
-    protected String getOAuthScope() {
-        return "user";
-    }
-
-    @Override
-    protected String getProfileUrl(final OAuth2AccessToken accessToken) {
-        return "https://api.foursquare.com/v2/users/self?v=20131118";
-    }
-
-    @Override
-    protected FoursquareProfile extractUserProfile(String body) throws HttpAction {
-        FoursquareProfile profile = getProfileDefinition().newProfile();
-        JsonNode json = JsonHelper.getFirstNode(body);
-        if (json == null) {
-            return profile;
-        }
-        JsonNode response = (JsonNode) JsonHelper.getElement(json, "response");
-        if (response == null) {
-            return profile;
-        }
-        JsonNode user = (JsonNode) JsonHelper.getElement(response, "user");
-        if (user != null) {
-            profile.setId(JsonHelper.getElement(user, "id"));
-
-            for (final String attribute : getProfileDefinition().getPrimaryAttributes()) {
-                getProfileDefinition().convertAndAdd(profile, attribute, JsonHelper.getElement(user, attribute));
-            }
-        }
-        return profile;
     }
 }