+++ pac4j-oauth/src/main/java/org/pac4j/oauth/client/PayPalClient.java
@@ -1,13 +1,7 @@
 package org.pac4j.oauth.client;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.github.scribejava.core.builder.api.BaseApi;
-import com.github.scribejava.core.model.OAuth2AccessToken;
-import com.github.scribejava.core.oauth.OAuth20Service;
 import org.pac4j.core.context.WebContext;
-import org.pac4j.core.exception.HttpAction;
 import org.pac4j.core.util.CommonHelper;
-import org.pac4j.oauth.profile.JsonHelper;
 import org.pac4j.oauth.profile.paypal.PayPalProfile;
 import org.pac4j.oauth.profile.paypal.PayPalProfileDefinition;
 import org.pac4j.scribe.builder.api.PayPalApi20;
@@ -23,63 +17,33 @@ import org.pac4j.scribe.builder.api.PayPalApi20;
  * @author Jerome Leleu
  * @since 1.4.2
  */
-public class PayPalClient extends BaseOAuth20Client<PayPalProfile> {
+public class PayPalClient extends OAuth20Client<PayPalProfile> {
     
     public final static String DEFAULT_SCOPE = "openid profile email address";
     
     protected String scope = DEFAULT_SCOPE;
     
     public PayPalClient() {
-        setTokenAsHeader(true);
     }
     
     public PayPalClient(final String key, final String secret) {
         setKey(key);
         setSecret(secret);
-        setTokenAsHeader(true);
     }
     
     @Override
     protected void internalInit(final WebContext context) {
         CommonHelper.assertNotBlank("scope", this.scope);
+        configuration.setApi(new PayPalApi20());
+        configuration.setProfileDefinition(new PayPalProfileDefinition());
+        configuration.setScope(this.scope);
+        configuration.setHasGrantType(true);
+        configuration.setTokenAsHeader(true);
+        setConfiguration(configuration);
+
         super.internalInit(context);
-        setProfileDefinition(new PayPalProfileDefinition());
     }
 
-    @Override
-    protected BaseApi<OAuth20Service> getApi() {
-        return new PayPalApi20();
-    }
-
-    @Override
-    protected String getOAuthScope() {
-        return this.scope;
-    }
-
-    @Override
-    protected  boolean hasOAuthGrantType() {
-        return true;
-    }
-
-    @Override
-    protected String getProfileUrl(final OAuth2AccessToken accessToken) {
-        return "https://api.paypal.com/v1/identity/openidconnect/userinfo?schema=openid";
-    }
-    
-    @Override
-    protected PayPalProfile extractUserProfile(final String body) throws HttpAction {
-        final PayPalProfile profile = getProfileDefinition().newProfile();
-        final JsonNode json = JsonHelper.getFirstNode(body);
-        if (json != null) {
-            final String userId = (String) JsonHelper.getElement(json, "user_id");
-            profile.setId(CommonHelper.substringAfter(userId, "/user/"));
-            for (final String attribute : getProfileDefinition().getPrimaryAttributes()) {
-                getProfileDefinition().convertAndAdd(profile, attribute, JsonHelper.getElement(json, attribute));
-            }
-        }
-        return profile;
-    }
-    
     public String getScope() {
         return this.scope;
     }